+++ packages/mint-components/src/components/sqm-big-stat/useBigStat.tsx
@@ -181,7 +181,7 @@ const referralsWeekQuery = (programId: string) => {
 
   const filter = {
     ...programFilter,
-    dateReferralStarted_timeframe: "this_month",
+    dateReferralStarted_timeframe: "this_week",
   };
   return debugQuery(
     gql`
@@ -383,6 +383,59 @@ const rewardsWeekQuery = (
   );
 };
 
+const rewardsPendingQuery = (
+  programId: string,
+  locale: string,
+  type: string,
+  unit: string,
+  global = ""
+) => {
+  return debugQuery(
+    gql`
+      query (
+        $programId: ID
+        $type: RewardType
+        $unit: String!
+        $locale: RSLocale
+      ) {
+        fallback: formatRewardPrettyValue(
+          value: 0
+          unit: $unit
+          locale: $locale
+          formatType: UNIT_FORMATTED
+        )
+        viewer: viewer {
+          ... on User {
+            rewardBalanceDetails(
+              programId: $programId
+              filter: { type_eq: $type, unit_eq: $unit }
+              locale: $locale
+            ) {
+              ... on CreditRewardBalance {
+                prettyPendingCredit
+              }
+            }
+          }
+        }
+      }
+    `,
+    {
+      programId: !global && programId !== "classic" ? programId : null,
+      type,
+      unit,
+      locale,
+    },
+    (res) => {
+      const arr = res.data?.viewer?.rewardBalanceDetails;
+      const fallback = res.data?.fallback;
+      return {
+        value: arr?.[0]?.prettyPendingCredit || 0,
+        statvalue: arr?.[0]?.prettyPendingCredit || fallback,
+      };
+    }
+  );
+};
+
 const rewardsRedeemedQuery = (
   programId: string,
   locale: string,
@@ -431,6 +484,156 @@ const rewardsRedeemedQuery = (
       return {
         value: arr?.[0]?.prettyRedeemedCredit || 0,
         statvalue: arr?.[0]?.prettyRedeemedCredit || fallback,
+      };
+    }
+  );
+};
+
+const rewardsRedeemedWeekQuery = (
+  programId: string,
+  locale: string,
+  type: string,
+  unit: string,
+  global = ""
+) => {
+  return debugQuery(
+    gql`
+      query ($programId: ID, $type: RewardType, $unit: String!) {
+        viewer: viewer {
+          ... on User {
+            rewards(
+              limit: 1000
+              filter: {
+                programId_eq: $programId
+                type_eq: $type
+                unit_eq: $unit
+                dateRedeemed_timeframe: "this_week"
+              }
+            ) {
+              data {
+                redeemedCredit
+              }
+              totalCount
+            }
+          }
+        }
+      }
+    `,
+    {
+      programId: !global && programId !== "classic" ? programId : null,
+      type,
+      unit,
+    },
+    (res) => {
+      const rewardData = res.data?.viewer?.rewards?.data;
+      const fallback = 0;
+      const redeemedTotal = rewardData?.reduce(
+        (total, reward) => (total += reward.redeemedCredit),
+        fallback
+      );
+
+      if (redeemedTotal !== undefined) {
+        const query = `query formatRewardPrettyValue(
+          $value: Int!
+          $unit: String!
+          $locale: RSLocale
+        ) {
+          formatRewardPrettyValue(
+          value: $value
+          unit: $unit
+          locale: $locale
+          formatType: UNIT_FORMATTED
+        )}`;
+
+        const result = useQuery(query, {
+          value: redeemedTotal,
+          unit,
+          locale,
+        });
+        return {
+          value: redeemedTotal || fallback,
+          statvalue: result?.data?.formatRewardPrettyValue,
+        };
+      }
+
+      return {
+        value: fallback,
+        statvalue: "...",
+      };
+    }
+  );
+};
+
+const rewardsRedeemedMonthQuery = (
+  programId: string,
+  locale: string,
+  type: string,
+  unit: string,
+  global = ""
+) => {
+  return debugQuery(
+    gql`
+      query ($programId: ID, $type: RewardType, $unit: String!) {
+        viewer: viewer {
+          ... on User {
+            rewards(
+              limit: 1000
+              filter: {
+                programId_eq: $programId
+                type_eq: $type
+                unit_eq: $unit
+                dateRedeemed_timeframe: "this_month"
+              }
+            ) {
+              data {
+                redeemedCredit
+              }
+              totalCount
+            }
+          }
+        }
+      }
+    `,
+    {
+      programId: !global && programId !== "classic" ? programId : null,
+      type,
+      unit,
+    },
+    (res) => {
+      const rewardData = res.data?.viewer?.rewards?.data;
+      const fallback = 0;
+      const redeemedTotal = rewardData?.reduce(
+        (total, reward) => (total += reward.redeemedCredit),
+        fallback
+      );
+
+      if (redeemedTotal !== undefined) {
+        const query = `query formatRewardPrettyValue(
+          $value: Int!
+          $unit: String!
+          $locale: RSLocale
+        ) {
+          formatRewardPrettyValue(
+          value: $value
+          unit: $unit
+          locale: $locale
+          formatType: UNIT_FORMATTED
+        )}`;
+
+        const result = useQuery(query, {
+          value: redeemedTotal,
+          unit,
+          locale,
+        });
+        return {
+          value: redeemedTotal || fallback,
+          statvalue: result?.data?.formatRewardPrettyValue,
+        };
+      }
+
+      return {
+        value: fallback,
+        statvalue: "...",
       };
     }
   );
@@ -603,6 +806,32 @@ const rewardsBalanceQuery = (
   );
 };
 
+const trafficQuery = () => {
+  return debugQuery(
+    gql`
+      query traffic {
+        viewer: viewer {
+          ... on User {
+            stats {
+              traffic
+            }
+          }
+        }
+      }
+    `,
+    {},
+    (res) => {
+      const traffic = res.data?.viewer?.stats.traffic;
+      const fallback = 0;
+
+      return {
+        value: traffic || fallback,
+        statvalue: traffic || fallback,
+      };
+    }
+  );
+};
+
 // functions are of the form (programId: string, ...args: string) => string
 export const queries: {
   [key: string]: {
@@ -617,9 +846,21 @@ export const queries: {
     label: "Rewards Earned",
     query: rewardsAssignedQuery,
   },
+  rewardsPending: {
+    label: "Rewards Pending",
+    query: rewardsPendingQuery,
+  },
   rewardsRedeemed: {
     label: "Rewards Paid",
     query: rewardsRedeemedQuery,
+  },
+  rewardsRedeemedMonth: {
+    label: "Rewards Paid - This Month",
+    query: rewardsRedeemedMonthQuery,
+  },
+  rewardsRedeemedWeek: {
+    label: "Rewards Paid - This Week",
+    query: rewardsRedeemedWeekQuery,
   },
   rewardsAvailable: {
     label: "Rewards Available",
@@ -669,6 +910,10 @@ export const queries: {
     label: "Custom Fields",
     query: customFieldsQuery,
   },
+  traffic: {
+    label: "Traffic",
+    query: trafficQuery,
+  },
 };
 
 // this should be exposed in documentation somehow
@@ -696,8 +941,20 @@ export const StatPaths = [
     route: "/(rewardsAssigned)/:statType/:unit/:global?",
   },
   {
+    name: "rewardsPending",
+    route: "/(rewardsPending)/:statType/:unit/:global?",
+  },
+  {
     name: "rewardsRedeemed",
     route: "/(rewardsRedeemed)/:statType/:unit/:global?",
+  },
+  {
+    name: "rewardsRedeemedWeek",
+    route: "/(rewardsRedeemedWeek)/:statType/:unit/:global?",
+  },
+  {
+    name: "rewardsRedeemedMonth",
+    route: "/(rewardsRedeemedMonth)/:statType/:unit/:global?",
   },
   {
     name: "rewardsAvailable",
@@ -707,6 +964,10 @@ export const StatPaths = [
     name: "rewardBalance",
     route:
       "/(rewardBalance)/:statType/:unit/:format([prettyValue|value]*)?/:global?",
+  },
+  {
+    name: "traffic",
+    route: "/(traffic)",
   },
 ];