+++ packages/paypal-components/src/global/global.ts
@@ -64,7 +64,7 @@ import {
   SlTooltip,
   setBasePath,
   registerIconLibrary,
-} from "@shoelace-style/shoelace";
+} from "@saasquatch/shoelace";
 
 try {
   setBasePath("/dist/shoelace");